Consistent naming for EIP patterns #733
