+++ eip-publish-subscribe/src/main/resources/logback.xml
