+++ pom.xml
@@ -117,11 +117,11 @@
         <module>business-delegate</module>
         <module>half-sync-half-async</module>
         <module>layers</module>
-        <module>message-channel</module>
+        <module>eip-message-channel</module>
         <module>fluentinterface</module>
         <module>reactor</module>
         <module>caching</module>
-        <module>publish-subscribe</module>
+        <module>eip-publish-subscribe</module>
         <module>delegation</module>
         <module>event-driven-architecture</module>
         <module>api-gateway</module>
