--- publish-subscribe/README.md
@@ -1,25 +1,0 @@
----
-layout: pattern
-title: Publish Subscribe
-folder: publish-subscribe
-permalink: /patterns/publish-subscribe/
-categories: Integration
-tags:
- - Java
- - EIP
- - Apache Camel™
----
-
-## Intent
-Broadcast messages from sender to all the interested receivers.
-
-![alt text](./etc/publish-subscribe.png "Publish Subscribe Channel")
-
-## Applicability
-Use the Publish Subscribe Channel pattern when
-
-* two or more applications need to communicate using a messaging system for broadcasts.
-
-## Credits
-
-* [J2EE Design Patterns](http://www.amazon.com/J2EE-Design-Patterns-William-Crawford/dp/0596004273/ref=sr_1_2)
